Fix overflow in Multiply when multiplying a number with a large fractional part by a large number. Original: 09afbf073e9c2321fb22445f290b6554d2a32fb5

+++ lal/packages/date/src/XLALTime.c
@@ -94,7 +94,7 @@ int XLALGPSCmp( const LIGOTimeGPS *t0, const LIGOTimeGPS *t1 )
 LIGOTimeGPS *XLALGPSMultiply( LIGOTimeGPS *gps, REAL8 x )
 {
   LIGOTimeGPS hi, mi;
-  INT8 lo = gps->gpsNanoSeconds * x + 0.5;
+  INT8 lo = ((INT8) gps->gpsNanoSeconds) * x + 0.5;
 
   XLALGPSSetREAL8(&mi, (gps->gpsSeconds % 512) * x);
   gps->gpsSeconds -= gps->gpsSeconds % 512;
